+++ Tests/DipTests/AutoInjectionTests.swift
@@ -339,6 +339,41 @@ class AutoInjectionTests: XCTestCase {
     let server = client.server
     XCTAssertTrue(client === server?.client)
   }
-  
+
+  func testThatItDoesNotAutoInjectIfDisabledInDefinition() {
+    container.register { ServerImp() as Server }
+    container.register { ClientImp() as Client }
+      .autoInjectingProperties(false)
+
+    let client = try! container.resolve() as Client
+    let server = client.server
+
+    XCTAssertNil(server)
+  }
+
+  func testThatItDoesNotAutoInjectIfDisabledInContainer() {
+    let container = DependencyContainer(autoInjectProperties: false)
+    container.register { ServerImp() as Server }
+    container.register { ClientImp() as Client }
+
+    let client = try! container.resolve() as Client
+    let server = client.server
+
+    XCTAssertNil(server)
+  }
+
+  func testThatItAutoInjectsWhenOverridenInDefinition() {
+    let container = DependencyContainer(autoInjectProperties: false)
+    container.register { ServerImp() as Server }
+    container.register { ClientImp() as Client }
+      .autoInjectingProperties(true)
+
+    let client = try! container.resolve() as Client
+    let server = client.server
+
+    XCTAssertNotNil(server)
+    XCTAssertNil(server?.client)
+  }
+
 }